Downgrade Java version from 17 to 1.8
The Java version used in the project has been downgraded from 17 to 1.8. This affects the `app`, `core`, and `favorite` modules.

The `compileOptions` and `kotlinOptions` in the `build.gradle.kts` files for these modules have been updated to reflect the new Java version.

Additionally, the Gradle JVM in `.idea/gradle.xml` has been set to `jbr-17` and the `gradleHome` path has been updated.