Modify recommended recipe's title and its behaviour
The recommended recipes title is changed from "Recommended Recipes" to "Give it a shot".
The title now also updates to "Search result" when the user performs a search, and displays "No recipe found for..." if no results are found.
Previously, the app would automatically search for "chicken" if no cached recipes were available. This behavior has been removed.